gummiboot: Add 32 bit compatible rdtsc asm
Gcc's inline asm constraints have different meanings on x86_64 and ia32.
Include a 32 bit version for the rdtsc function. Drop the empty 32 bit
version of time_usec as it and the cpuid function both function properly
when compiled for 32 bit systems.

Tested on the following CPU:
Intel(R) Atom(TM) CPU E640   @ 1.00GHz

A value of 1000000000 was detected.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>
Cc: "Ylinen, Mikko" <mikko.ylinen@intel.com>