Autodetect both x64 and ia32 boot*.efi payloads
The EFI specification documents /EFI/BOOT/bootx64.efi for x86_64
machines and /EFI/BOOT/bootia32.efi for ia32 machines. Update the auto
detection to allow for both.

Add the MACHINE_TYPE_NAME define to the efi/src build so we can use it
as the standard suffix for the EFI payloads (ia32 or x64).

Signed-off-by: Darren Hart <dvhart@linux.intel.com>