Revert "obsoleted"
This reverts commit 55df1539c9d330732e88bd196afee386db6e4a1d.

This fork is now being used for Solus (and anyone else if they're interested)
because we need the full features of gummiboot, outside of the context of
systemd.

It's been stated on several occassions that systemd-boot is "not a direct
successor", and "systemd-boot does not support some of the features gummiboot
supported". As such, this fork becomes a separate and independent bootloader.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>