setup: Ensure we remove a conflicting target file in copy_file
Due to the oddities of FAT such as case-ignorance, goofiboot actually
looks for the case-*compatible* path before updating the ESP. This means
that the path we look for mightn't be the path we think we're renaming to.

To counteract this, and to be a good FAT citizen, we'll remove the target
file in this stage of the copy, prior to the rename. Also of note is that
FAT in of itself is delicate and you shouldn't rename and obliterate existing
files. In this case we always remove the target, and then rename to it, with
sync()'s in between to avoid corruption to the ESP.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>